requirements.txt
kubernetes>=10.0.1
powerline-status>=2.7
